+++ src/help_functions.cu
@@ -6,7 +6,7 @@ extern "C" {
 __global__ void sum_reduction(
 	conjugrad_float_t *d_in,
 	conjugrad_float_t *d_out,
-	int nvar
+    unsigned long nvar
 ) {
 	int i = threadIdx.x;
